~ Completely different API! Zustand meets Redux meets Pinia
This is a complete change to the API. I got fed up of dealing with
classes and `this` 😱 Much more fun to play with functional! This is
essentially Zustand but how _I_ would have built it 🤓 Immer is built in,
actions are called events (as they should be), and selectors are moved
to the store. This is going out as a patch, living next to the original
2n8 until I've tested it in projects.